openclaw-backend/requirements.txt
fastapi
uvicorn
python-multipart
websockets
psutil
aiofiles
aiohttp
requests
llama-cpp-python
openai-whisper
sentence-transformers
chromadb
torch
numpy
scipy
